requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
click==8.0.1
discord.py==1.7.2
Flask==2.0.1
itsdangerous==2.0.1
multidict==5.1.0
Werkzeug==2.0.1
yarl==1.6.3
